fix: decouple openchoreo-clent-node from Backstage dependencies
  Extract a standalone Logger interface and move the Backstage-specific
  createOpenChoreoClientFromConfig factory into a dedicated backstage.ts
  module, so the core client can be consumed by non-Backstage projects
  (e.g. the VSCode extension). Backstage packages are now optional peer
  dependencies.
