+++ packages/openchoreo-client-node/package.json
@@ -26,10 +26,20 @@
     "postpack": "backstage-cli package postpack"
   },
   "dependencies": {
-    "@backstage/backend-plugin-api": "1.4.3",
-    "@backstage/config": "1.3.4",
     "json-schema": "0.4.0",
     "openapi-fetch": "0.13.8"
+  },
+  "peerDependencies": {
+    "@backstage/backend-plugin-api": ">=1.0.0",
+    "@backstage/config": ">=1.0.0"
+  },
+  "peerDependenciesMeta": {
+    "@backstage/backend-plugin-api": {
+      "optional": true
+    },
+    "@backstage/config": {
+      "optional": true
+    }
   },
   "devDependencies": {
     "@backstage/cli": "0.34.3",
